# Exclude annotation sources if not specified in protobuf for improved api query_variant_calls performance.