--- a/src/main/cpp/src/genomicsdb/genomicsdb_iterators.cc
+++ b/src/main/cpp/src/genomicsdb/genomicsdb_iterators.cc
@@ -523,7 +523,6 @@ bool SingleCellTileDBIterator::advance_to_next_useful_cell(const uint64_t min_nu
       if (genomicsdb_columnar_field.is_variable_length_field()) {
         buffers.push_back(const_cast<GenomicsDBBuffer *>(buffer.first)->get_offsets_pointer());
         buffer_sizes.push_back(buffer.first->get_offsets_size_in_bytes());
-        positions.push_back((int64_t)buffer.second);
       }
       buffers.push_back(const_cast<uint8_t *>(buffer.first->get_raw_pointer()));
       buffer_sizes.push_back(buffer.first->get_data_vector_size_in_bytes());