- Exclude annotation sources if not specified in protobuf for improved api `query_variant_calls` performance.
- Bug fixes and added more tests for query filter expressions.
- Added a test case for the GenomicsDB demo test case. This can be run as ` GENOMICSDB_DEMO_WS=~/genomicsdb_demo/ws ./api_tests [genomicsdb_demo]`. The test case outputs times for each of the runs with and without filters. The test checks the count of samples for the interval `17:7571719-7590868` with the most restrictive filter `"REF=="A" && ALT|="T" && GT &= "1/1"` running under 3s using 0.3G memory.
- Also added Java test example for the demo. Can be invoked as `java -cp genomicsdb-1.5.1-SNAPSHOT-allinone-spark.jar TestGenomicsDBDemo.java`

See TileDB related changes - https://github.com/OmicsDataAutomation/TileDB/pull/153

C++ tests on MacOS M2 with 32GB only counts - `./api_tests [genomicsdb_demo]`
```
Elapsed Time=**1s**
Elapsed Time=**2s** for REF=="A"
Elapsed Time=**2s** for REF=="A" && ALT|="T"
Elapsed Time=**2s** for REF=="A" && ALT|="T" && GT &= "1/1"
```

Java tests on MacOS M2 with 32GB - ` java -cp genomicsdb-1.5.1-SNAPSHOT-allinone-spark.jar TestGenomicsDBDemo.java `
```
Summary: Elapsed time: **188s**
Interval: 2507743582-2507762731 calls=2962039
Summary for REF=="A": Elapsed time: **33s**
Interval: 2507743582-2507762731 calls=400432
Summary for REF=="A" && ALT|="T": Elapsed time: **7s**
Interval: 2507743582-2507762731 calls=82245
Summary for REF=="A" && ALT|="T" && GT&="1/1": Elapsed time: **8s**
Interval: 2507743582-2507762731 calls=82245
```

C++ tests on Centos7(256G) only counts - `./api_tests [genomicsdb_demo]`
```
Elapsed Time=**4s** for 
Elapsed Time=**6s** for REF=="A"
Elapsed Time=**7s** for REF=="A" && ALT|="T"
Elapsed Time=**9s** for REF=="A" && ALT|="T" && GT &= "1/1"
```

Java tests on Centos 7(256G) - `/usr/bin/time java -cp target/genomicsdb-1.5.1-SNAPSHOT-allinone-spark.jar TestGenomicsDBDemo.java
```
Summary for REF=="A" && ALT|="T" && GT&="1/1": Elapsed time: **32s**
Interval: 2507743582-2507762731 calls=82245
36.75user 1.34system 0:33.83elapsed 112%CPU (0avgtext+0avgdata 721000maxresident)k
0inputs+46952outputs (0major+171066minor)pagefaults 0swaps
```

Java tests on Centos 7(256G) - `/usr/bin/time java -cp target/genomicsdb-1.5.1-SNAPSHOT-allinone-spark.jar  -Xms1G -Xmx1G TestGenomicsDBDemo.java`
```
Summary for REF=="A" && ALT|="T" && GT&="1/1": Elapsed time: **31s**
Interval: 2507743582-2507762731 calls=82245
35.46user 1.41system 0:32.60elapsed 113%CPU (0avgtext+0avgdata 671916maxresident)k
0inputs+46912outputs (0major+180733minor)pagefaults 0swaps
```